feat: 인증 확인해서 알맞는 라우팅으로 가게 세팅
