+++ src/app/page.tsx
@@ -1,11 +1,3 @@
 export default function Home() {
-  return (
-    <main style={{ padding: '2rem', textAlign: 'center' }}>
-      <h1>Horizon</h1>
-      <p>Welcome to Horizon Application</p>
-      <a href="/signin" style={{ color: 'blue', textDecoration: 'underline' }}>
-        Sign In
-      </a>
-    </main>
-  );
+  return <> </>;
 }
